+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/dao/PermissionType.java
@@ -1,0 +1,10 @@
+package com.elmakers.mine.bukkit.plugins.persistence.dao;
+
+// Enums don't need annotations to be persistable
+public enum PermissionType
+{
+	DEFAULT,
+	ALLOW_ALL,
+	OPS_ONLY,
+	PLAYER_ONLY
+};