+++ PersistencePlugin/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+# 0.27
+ - Don't persist command.children, it can be built at runtime from the parents.
+ - Permissions integration (Permissions.hasPermission, and CommandData.permissionNode)
+ - Remove NetherGate-specific stuff from WorldData- my bad! I promise not to abuse my dual dev/client status in the future :)
+
 # 0.26
  
   - Prefix contained value names